requirements.txt
# Core dependencies
Pillow>=9.5.0
numpy>=1.24.0
PyMuPDF>=1.22.5
PyPDF2>=3.0.0
pypdfium2>=4.16.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
# Add any additional dependencies for NLP, image processing, etc. 